// CorporateQnA/Client App/src/app/shared/models/question-info.model.ts
export class QuestionInfo {
    id: number;
    askedById: number;
    askedBy: string;
    profileImage: string;
    title: string;
    description: string;
    categoryId:number;
    upvotes: number;
    totalViews: number;
    totalAnswers: number;
    askedOn: string;
    resolved: boolean;
    status:number;
}